Tweaked a bunch of things and started styling.
