+++ src/app/loadout-drawer/subclass-drawer/Abilities.m.scss
@@ -1,14 +1,34 @@
+$title-color: #979797;
+
+.container {
+  display: flex;
+  flex-direction: column;
+  width: fit-content;
+}
+
+.title {
+  border-bottom: 1px solid $title-color;
+  font-weight: bold;
+  color: $title-color;
+  margin-bottom: 4px;
+}
+
 .abilities {
   display: flex;
   flex-direction: row;
+  margin-bottom: 16px;
 }
 
-.ability {
+.options {
   display: flex;
-  flex-direction: row;
+  flex-direction: column;
   margin-right: 16px;
-}
 
-.option {
-  margin-right: 4px;
+  &:last-child {
+    margin-right: 0;
+  }
+
+  > * {
+    margin-bottom: 4px;
+  }
 }
