+++ src/app/loadout-drawer/subclass-drawer/Abilities.tsx
@@ -1,7 +1,7 @@
-import { isPluggableItem } from 'app/inventory/store/sockets';
-import { DestinyInventoryItemDefinition } from 'bungie-api-ts/destiny2';
+import { PluggableInventoryItemDefinition } from 'app/inventory/item-types';
 import React from 'react';
-import SocketOptions from './SocketOptions';
+import styles from './Abilities.m.scss';
+import Option from './Option';
 import { SelectedPlugs, SocketWithOptions } from './types';
 
 export default function Abilities({
@@ -13,18 +13,31 @@ export default function Abilities({
   selectedPlugs: SelectedPlugs;
   setSelectedPlugs(selectedPlugs: SelectedPlugs): void;
 }) {
-  const selectAbility = (ability: DestinyInventoryItemDefinition) => {
-    if (isPluggableItem(ability)) {
-      const { plugCategoryHash } = ability.plug;
-      setSelectedPlugs({ ...selectedPlugs, [plugCategoryHash]: [ability] });
-    }
+  const selectAbility = (ability: PluggableInventoryItemDefinition) => {
+    const { plugCategoryHash } = ability.plug;
+    setSelectedPlugs({ ...selectedPlugs, [plugCategoryHash]: [ability] });
   };
   return (
-    <SocketOptions
-      socketsWithOptions={abilities}
-      direction="column"
-      selectedPlugs={selectedPlugs}
-      onSelect={selectAbility}
-    />
+    <div className={styles.container}>
+      {Boolean(abilities.length) && <div className={styles.title}>{abilities[0].title}</div>}
+      <div className={styles.abilities}>
+        {abilities.map(({ socket, options }) => (
+          <div key={socket.plugged?.plugDef.itemTypeDisplayName} className={styles.options}>
+            {options.map((option) => (
+              <Option
+                key={option.hash}
+                option={option}
+                isSelected={plugIsSelected(selectedPlugs, option)}
+                onSelect={() => selectAbility(option)}
+              />
+            ))}
+          </div>
+        ))}
+      </div>
+    </div>
   );
+}
+
+function plugIsSelected(selectedPlugs: SelectedPlugs, option: PluggableInventoryItemDefinition) {
+  return Boolean(selectedPlugs[option.plug.plugCategoryHash]?.includes(option));
 }
