+++ src/app/loadout-drawer/subclass-drawer/Option.m.scss
@@ -1,0 +1,10 @@
+@import '../../variables';
+
+.option {
+  cursor: pointer;
+}
+
+.selected {
+  cursor: default;
+  outline: 2px solid $orange !important;
+}
