+++ src/app/loadout-drawer/subclass-drawer/Option.m.scss.d.ts
@@ -1,11 +1,8 @@
 // This file is automatically generated.
 // Please do not change this file!
 interface CssExports {
-  'column': string;
   'option': string;
-  'options': string;
   'selected': string;
-  'socketsWithOptions': string;
 }
 export const cssExports: CssExports;
 export default cssExports;
