--- src/app/loadout-drawer/subclass-drawer/SocketOptions.m.scss
@@ -1,26 +1,0 @@
-@import '../../variables';
-
-.socketsWithOptions {
-  display: flex;
-  flex-direction: row;
-  margin-bottom: 16px;
-}
-
-.options {
-  display: flex;
-  margin-right: 16px;
-}
-
-.column {
-  flex-direction: column;
-}
-
-.option {
-  margin: 0 4px 4px 0;
-  cursor: pointer;
-}
-
-.selected {
-  cursor: default;
-  outline: 2px solid $orange !important;
-}
