+++ src/app/loadout-drawer/subclass-drawer/SubclassDrawer.m.scss.d.ts
@@ -1,13 +1,10 @@
 // This file is automatically generated.
 // Please do not change this file!
 interface CssExports {
-  'abilities': string;
-  'aspects': string;
   'container': string;
-  'fragments': string;
-  'optionsGrid': string;
+  'selected': string;
+  'subclass': string;
   'subclasses': string;
-  'super': string;
 }
 export const cssExports: CssExports;
 export default cssExports;
