+++ src/app/loadout-drawer/subclass-drawer/SubclassDrawer.tsx
@@ -1,25 +1,19 @@
-import { D2ManifestDefinitions } from 'app/destiny2/d2-definitions';
 import { bungieNetPath } from 'app/dim-ui/BungieImage';
 import ClassIcon from 'app/dim-ui/ClassIcon';
 import Sheet from 'app/dim-ui/Sheet';
 import ConnectedInventoryItem from 'app/inventory/ConnectedInventoryItem';
 import { DimItem, PluggableInventoryItemDefinition } from 'app/inventory/item-types';
 import { allItemsSelector } from 'app/inventory/selectors';
-import { isPluggableItem } from 'app/inventory/store/sockets';
 import { useD2Definitions } from 'app/manifest/selectors';
 import { useIsPhonePortrait } from 'app/shell/selectors';
 import { DestinyClass } from 'bungie-api-ts/destiny2';
-import { SocketCategoryHashes } from 'data/d2/generated-enums';
+import clsx from 'clsx';
 import _ from 'lodash';
 import React, { useMemo, useState } from 'react';
-import ReactDOM from 'react-dom';
 import { shallowEqual, useSelector } from 'react-redux';
-import Abilities from './Abilities';
-import AspectAndFragmentDrawer from './AspectAndFragmentDrawer';
-import Aspects from './Aspects';
-import SocketOptions from './SocketOptions';
 import styles from './SubclassDrawer.m.scss';
-import { SelectedPlugs, SocketWithOptions } from './types';
+import SubclassOptions from './SubclassOptions';
+import { SelectedPlugs } from './types';
 
 export default function SubclassDrawer({
   classType,
@@ -71,7 +65,9 @@ export default function SubclassDrawer({
             <div
               key={subclass.id}
               onClick={() => setSelectedSubclass(subclass)}
-              className="loadout-item"
+              className={clsx('loadout-item', styles.subclass, {
+                [styles.selected]: subclass.id === selectedSubclass?.id,
+              })}
             >
               <ConnectedInventoryItem item={subclass} ignoreSelectedPerks={true} />
               {subclass.type === 'Class' && (
@@ -90,121 +86,5 @@ export default function SubclassDrawer({
         )}
       </div>
     </Sheet>
-  );
-}
-
-function getSocketsWithOptionsForCategory(
-  defs: D2ManifestDefinitions,
-  item: DimItem,
-  categoryHash: SocketCategoryHashes
-) {
-  const rtn: SocketWithOptions[] = [];
-  const indexes =
-    item.sockets?.categories.find((category) => category.category.hash === categoryHash)
-      ?.socketIndexes || [];
-  for (const index of indexes) {
-    const socket = item.sockets?.allSockets[index];
-    const plugSetHash = socket?.socketDefinition.reusablePlugSetHash;
-
-    // Non-super case
-    if (plugSetHash) {
-      const plugSet = plugSetHash !== undefined ? defs.PlugSet.get(plugSetHash) : undefined;
-      const options = plugSet?.reusablePlugItems
-        ?.map((plugItem) => defs.InventoryItem.get(plugItem.plugItemHash))
-        .filter(isPluggableItem);
-
-      if (socket && options?.length) {
-        rtn.push({ socket, options });
-      }
-    } else if (socket) {
-      const initialItemHash = socket?.socketDefinition.singleInitialItemHash;
-      const initialItem = defs.InventoryItem.get(initialItemHash);
-      rtn.push({ socket, options: [initialItem] });
-    }
-  }
-  return rtn;
-}
-
-function SubclassOptions({
-  selectedSubclass,
-  defs,
-  selectedPlugs,
-  setSelectedPlugs,
-}: {
-  selectedSubclass: DimItem;
-  defs: D2ManifestDefinitions;
-  selectedPlugs: SelectedPlugs;
-  setSelectedPlugs(selectedPlugs: SelectedPlugs): void;
-}) {
-  const [showPlugPicker, setShowPlugPicker] = useState(false);
-
-  const abilities = getSocketsWithOptionsForCategory(
-    defs,
-    selectedSubclass,
-    SocketCategoryHashes.Abilities
-  );
-  const supers = getSocketsWithOptionsForCategory(
-    defs,
-    selectedSubclass,
-    SocketCategoryHashes.Super
-  );
-  const aspects = getSocketsWithOptionsForCategory(
-    defs,
-    selectedSubclass,
-    SocketCategoryHashes.Aspects
-  );
-  const fragments = getSocketsWithOptionsForCategory(
-    defs,
-    selectedSubclass,
-    SocketCategoryHashes.Fragments
-  );
-
-  return (
-    <div className={styles.optionsGrid}>
-      <div className={styles.super}>
-        <SocketOptions socketsWithOptions={supers} direction="row" selectedPlugs={selectedPlugs} />
-      </div>
-      <div className={styles.abilities}>
-        <Abilities
-          abilities={abilities}
-          selectedPlugs={selectedPlugs}
-          setSelectedPlugs={setSelectedPlugs}
-        />
-      </div>
-      <div className={styles.aspects}>
-        <Aspects
-          aspects={aspects}
-          selectedPlugs={selectedPlugs}
-          setSelectedPlugs={setSelectedPlugs}
-          onOpenPlugPicker={() => setShowPlugPicker(true)}
-        />
-      </div>
-      <div className={styles.fragments}>
-        <Aspects
-          aspects={fragments}
-          selectedPlugs={selectedPlugs}
-          setSelectedPlugs={setSelectedPlugs}
-          onOpenPlugPicker={() => setShowPlugPicker(true)}
-        />
-      </div>
-      {showPlugPicker &&
-        ReactDOM.createPortal(
-          <AspectAndFragmentDrawer
-            aspects={aspects}
-            fragments={fragments}
-            selectedPlugs={selectedPlugs}
-            onAccept={(selected) => {
-              const groupedPlugs = _.groupBy(selected, (plug) => plug.plug.plugCategoryHash);
-              const newPlugs = { ...selectedPlugs };
-              for (const [plugCategoryHash, plugs] of Object.entries(groupedPlugs)) {
-                newPlugs[plugCategoryHash] = plugs;
-              }
-              setSelectedPlugs(newPlugs);
-            }}
-            onClose={() => setShowPlugPicker(false)}
-          />,
-          document.body
-        )}
-    </div>
   );
 }
