+++ src/app/loadout-drawer/subclass-drawer/SubclassOptions.m.scss
@@ -1,0 +1,25 @@
+.optionsGrid {
+  display: grid;
+  grid-template-columns: auto auto auto;
+  grid-template-rows: auto auto;
+}
+
+.super {
+  grid-column: 1;
+  grid-row: 1 / span 2;
+}
+
+.abilities {
+  grid-column: 2;
+  grid-row: 1;
+}
+
+.aspects {
+  grid-column: 3;
+  grid-row: 1;
+}
+
+.fragments {
+  grid-column: 2 / span 3;
+  grid-row: 2;
+}
