+++ src/app/loadout-drawer/subclass-drawer/SubclassOptions.m.scss.d.ts
@@ -1,9 +1,11 @@
 // This file is automatically generated.
 // Please do not change this file!
 interface CssExports {
-  'container': string;
-  'selectedAspects': string;
-  'title': string;
+  'abilities': string;
+  'aspects': string;
+  'fragments': string;
+  'optionsGrid': string;
+  'super': string;
 }
 export const cssExports: CssExports;
 export default cssExports;
